refactor(router): Update Router to be `providedIn: 'root'`
This commit updates the Router itself to be `providedIn: 'root'` with a
factory function rather than provided in the `RouterModule`.
