+++ packages/core/test/bundling/router/bundle.golden_symbols.json
@@ -1839,6 +1839,9 @@
     "name": "setUpAttributes"
   },
   {
+    "name": "setupRouter"
+  },
+  {
     "name": "shallowEqual"
   },
   {
